refactoring + fmt
