+++ tracker/Cargo.toml
@@ -9,5 +9,7 @@ edition = "2018"
 [dependencies]
 bitcoin = "0.26.1"
 bitcoincore-rpc = "0.13.0"
+
+[dev-dependencies]
 serde = "1.0.126"
 serde_json = "1.0.64"
