+++ tracker/src/index.rs
@@ -1,11 +1,10 @@
 use crate::bitcoin_client::BitcoinClient;
 use bitcoin::blockdata::opcodes;
 use bitcoin::blockdata::script::Instruction;
-use bitcoin::{Block, BlockHash, Transaction, Txid};
+use bitcoin::{BlockHash, Transaction, Txid};
 use bitcoincore_rpc::json::GetBlockHeaderResult;
 use std::cmp::Ordering;
 use std::collections::HashMap;
-use std::sync::mpsc::Receiver;
 use std::convert::TryInto;
 
 type BagId = [u8; 32];
@@ -61,8 +60,7 @@ impl<C: BitcoinClient> Index<C> {
                         }
                         let fork_position_in_checked_chain =
                             self.checked_chain.len() - reorg_info.discarded_blocks.len();
-                        self.checked_chain
-                            .drain(..fork_position_in_checked_chain);
+                        self.checked_chain.drain(..fork_position_in_checked_chain);
                         reorg_info.height_when_fork
                     }
                     None => self.height,
@@ -117,7 +115,10 @@ impl<C: BitcoinClient> Index<C> {
         let block = self.client.get_block(&hash).unwrap();
         let txs = block.txdata;
 
-        let mint_txs = txs.into_iter().filter_map(|tx| self.parse_mint_transaction(tx)).collect();
+        let mint_txs = txs
+            .into_iter()
+            .filter_map(|tx| self.parse_mint_transaction(tx))
+            .collect();
 
         mint_txs
     }
@@ -194,6 +195,7 @@ pub struct BitcoinMintOutputIndex {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::test_utils::*;
     use bitcoin::blockdata::script;
     use bitcoin::hashes::sha256d;
     use bitcoin::hashes::Hash;
@@ -202,103 +204,6 @@ mod tests {
     use std::cell::RefCell;
     use std::convert::Infallible;
     use std::rc::Rc;
-
-    #[derive(Clone)]
-    struct TestBlock {
-        height: u64,
-        block_hash: BlockHash,
-        in_main_chain: bool,
-        txs: Vec<Transaction>,
-    }
-
-    struct TestBitcoinClient {
-        blocks: Rc<RefCell<Vec<TestBlock>>>,
-    }
-
-    impl BitcoinClient for TestBitcoinClient {
-        type Err = Infallible;
-
-        fn get_blockchain_info(&self) -> Result<GetBlockchainInfoResult, Self::Err> {
-            let blocks = self.blocks.borrow();
-            Ok(GetBlockchainInfoResult {
-                chain: "rusttest".to_string(),
-                blocks: blocks.last().unwrap().height,
-                headers: 0,
-                best_block_hash: blocks.last().unwrap().block_hash.clone(),
-                difficulty: 0.0,
-                median_time: 0,
-                verification_progress: 0.0,
-                initial_block_download: false,
-                chain_work: vec![],
-                size_on_disk: 0,
-                pruned: false,
-                prune_height: None,
-                automatic_pruning: None,
-                prune_target_size: None,
-                softforks: Default::default(),
-                warnings: "".to_string(),
-            })
-        }
-
-        fn get_block_hash(&self, height: u64) -> Result<BlockHash, Self::Err> {
-            Ok(self
-                .blocks
-                .borrow()
-                .iter()
-                .find(|block| block.height == height && block.in_main_chain)
-                .unwrap()
-                .block_hash)
-        }
-
-        fn get_block_header_info(
-            &self,
-            hash: &BlockHash,
-        ) -> Result<GetBlockHeaderResult, Self::Err> {
-            let blocks = self.blocks.borrow();
-            let index = blocks
-                .iter()
-                .position(|block| block.block_hash == *hash)
-                .unwrap();
-            let block = &blocks[index];
-            Ok(GetBlockHeaderResult {
-                hash: block.block_hash.clone(),
-                confirmations: if block.in_main_chain { 1 } else { -1 },
-                height: block.height as usize,
-                version: 0,
-                version_hex: None,
-                merkle_root: Default::default(),
-                time: 0,
-                median_time: None,
-                nonce: 0,
-                bits: "".to_string(),
-                difficulty: 0.0,
-                chainwork: vec![],
-                n_tx: 0,
-                previous_block_hash: None,
-                next_block_hash: None,
-            })
-        }
-
-        fn get_block(&self, hash: &BlockHash) -> Result<Block, Self::Err> {
-            let blocks = self.blocks.borrow();
-            let index = blocks
-                .iter()
-                .position(|block| block.block_hash == *hash)
-                .unwrap();
-            let block = &blocks[index];
-            Ok(Block {
-                header: BlockHeader {
-                    version: 0,
-                    prev_blockhash: Default::default(),
-                    merkle_root: Default::default(),
-                    time: 0,
-                    bits: 0,
-                    nonce: 0,
-                },
-                txdata: block.txs.clone(),
-            })
-        }
-    }
 
     #[test]
     fn test_new_blocks() {
@@ -340,8 +245,10 @@ mod tests {
         index.check_last_blocks();
 
         assert_eq!(index.index.len(), 1);
+
         let txs_in_index = index.index.get(&block2.block_hash).unwrap();
         let block2_tx_out = &txs_in_index[0];
+
         assert_eq!(block2_tx_out.amount, 10);
         assert_eq!(block2_tx_out.bag_id, [1; 32]);
     }
@@ -353,6 +260,18 @@ mod tests {
         let forked_block = create_test_block_with_mint_tx(1, [3], [2; 32]);
         let forked_block2 = create_test_block_with_mint_tx(2, [4], [3; 32]);
 
+        let initial_blocks = vec![initial_block.clone()];
+        let blocks_chain_1 = vec![initial_block.clone(), block2.clone()];
+        let blocks_chain_2 = vec![
+            initial_block.clone(), // first block in both chains
+            TestBlock {
+                in_main_chain: false, // was in the main chain, after reorg is not
+                ..block2
+            },
+            forked_block.clone(),
+            forked_block2.clone(),
+        ];
+
         let blocks = Rc::new(RefCell::new(vec![initial_block.clone()]));
         let client = TestBitcoinClient {
             blocks: blocks.clone(),
@@ -363,21 +282,12 @@ mod tests {
         index.add_bag([2; 32]);
         index.add_bag([3; 32]);
 
-        blocks.borrow_mut().push(block2.clone());
+        *blocks.borrow_mut() = blocks_chain_1;
         index.check_last_blocks();
 
         assert_eq!(index.index.len(), 1);
 
-        *blocks.borrow_mut() = vec![
-            initial_block.clone(), // first block in both chains
-            TestBlock {
-                in_main_chain: false,
-                ..block2
-            }, // was in the main chain, after reorg is not
-            forked_block.clone(),
-            forked_block2.clone(),
-        ];
-
+        *blocks.borrow_mut() = blocks_chain_2;
         index.check_last_blocks();
 
         assert_eq!(index.index.len(), 2);
@@ -389,38 +299,5 @@ mod tests {
         let txs_in_index = index.index.get(&forked_block2.block_hash).unwrap();
         let forked_block2_tx_out = &txs_in_index[0];
         assert_eq!(forked_block2_tx_out.bag_id, [3; 32]);
-    }
-
-    fn create_test_block(height: u64, data: impl AsRef<[u8]>) -> TestBlock {
-        TestBlock {
-            height,
-            block_hash: BlockHash::from_hash(sha256d::Hash::hash(data.as_ref())),
-            in_main_chain: true,
-            txs: vec![],
-        }
-    }
-
-    fn create_test_block_with_mint_tx(
-        height: u64,
-        data: impl AsRef<[u8]>,
-        tx_data: impl AsRef<[u8]>,
-    ) -> TestBlock {
-        TestBlock {
-            height,
-            block_hash: BlockHash::from_hash(sha256d::Hash::hash(data.as_ref())),
-            in_main_chain: true,
-            txs: vec![Transaction {
-                version: 0,
-                lock_time: 0,
-                input: vec![],
-                output: vec![TxOut {
-                    value: 10,
-                    script_pubkey: script::Builder::new()
-                        .push_opcode(opcodes::all::OP_RETURN)
-                        .push_slice(tx_data.as_ref())
-                        .into_script(),
-                }],
-            }],
-        }
     }
 }
