+++ tracker/src/lib.rs
@@ -1,4 +1,7 @@
-pub mod index;
 pub mod bitcoin_client;
+pub mod index;
 
 pub use index::Index;
+
+#[cfg(test)]
+pub mod test_utils;
