+++ tracker/tests/regtest_bitcoin_node.rs
@@ -1,69 +1,30 @@
-use bitcoincore_rpc::{Client, Auth, RpcApi};
-use std::collections::HashMap;
-use bitcoin::{TxOut, Transaction, TxIn, Block, BlockHeader, BlockHash};
-use bitcoin::blockdata::{script, opcodes};
-use bitcoin::consensus::{Encodable, Decodable};
-use std::time::Duration;
+use bitcoin::blockdata::script;
+use bitcoin::consensus::Encodable;
 use bitcoin::hashes::hex::ToHex;
-use tracker::Index;
-use std::process::Stdio;
-use serde::Serialize;
-use serde_json::Value;
 use bitcoin::hashes::sha256d::Hash;
+use bitcoin::{BlockHash, Transaction, TxOut};
+use bitcoincore_rpc::{Auth, Client, RpcApi};
+use serde_json::Value;
+use std::process::Stdio;
 use std::str::FromStr;
-
-fn setup_node(port: u32, rpcport: u32, datadir: &str) -> std::process::Child {
-    let child = std::process::Command::new("bash")
-        .arg("./tests/setup_single_node.sh")
-        .arg(port.to_string())
-        .arg(rpcport.to_string())
-        .arg(datadir)
-        .stdout(Stdio::null())
-        .spawn()
-        .unwrap();
-    child
-}
-
-struct TempDir {
-    path: String,
-}
-
-impl TempDir {
-    fn new(path: String) -> Self {
-        std::fs::create_dir(path.as_str()).unwrap_or_else(|_| {
-            std::fs::remove_dir_all(path.as_str()).unwrap();
-            std::fs::create_dir(path.as_str()).unwrap();
-        });
-        Self {
-            path
-        }
-    }
-}
-
-impl Drop for TempDir {
-    fn drop(&mut self) {
-        std::fs::remove_dir_all(self.path.as_str()).unwrap_or_else(|_|());
-    }
-}
+use std::time::Duration;
+use tracker::Index;
 
 #[test]
 fn test_new_blocks_with_mint_txs() {
     const DIR: &'static str = "/tmp/bitcoin_test_node";
-    let dir = TempDir::new(DIR.to_string());
-    let node = setup_node(18444, 12001, DIR);
+    let _dir = TempDir::new(DIR.to_string());
+    let _node = setup_bitcoin_node(18444, 12001, DIR);
     std::thread::sleep(Duration::from_secs(1)); // Wait while node will started
     let client = Client::new(
         "http://localhost:12001".to_string(),
-        Auth::UserPass("rt".to_string(), "rt".to_string())
-    ).unwrap();
+        Auth::UserPass("rt".to_string(), "rt".to_string()),
+    )
+    .unwrap();
 
-    client.create_wallet(
-        "testwal",
-        None,
-        None,
-        None,
-        None,
-    ).unwrap();
+    client
+        .create_wallet("testwal", None, None, None, None)
+        .unwrap();
 
     let address = client.get_new_address(Some("testwal"), None).unwrap();
     client.generate_to_address(120, &address).unwrap();
@@ -74,7 +35,7 @@ fn test_new_blocks_with_mint_txs() {
         input: vec![],
         output: vec![TxOut {
             value: 10,
-            script_pubkey: script::Script::new_op_return(&[1; 32])
+            script_pubkey: script::Script::new_op_return(&[1; 32]),
         }],
     };
 
@@ -82,25 +43,28 @@ fn test_new_blocks_with_mint_txs() {
     consensus_encode_tx(&tx, &mut bytes).unwrap();
 
     let funded = client.fund_raw_transaction(&bytes, None, None).unwrap();
-    let signed = client.sign_raw_transaction_with_wallet(&funded.hex, None, None).unwrap();
+    let signed = client
+        .sign_raw_transaction_with_wallet(&funded.hex, None, None)
+        .unwrap();
     assert!(signed.complete);
     let tx_id = client.send_raw_transaction(&signed.hex).unwrap();
 
-    let value: Value = client.call(
-        "generateblock",
-        &[
-            Value::String(address.to_string()),
-            Value::Array(vec![Value::String(tx_id.as_hash().to_hex())])
-        ]).unwrap();
+    let value: Value = client
+        .call(
+            "generateblock",
+            &[
+                Value::String(address.to_string()),
+                Value::Array(vec![Value::String(tx_id.as_hash().to_hex())]),
+            ],
+        )
+        .unwrap();
 
     let hash = match value {
-        Value::Object(m) => {
-            match m.get("hash").unwrap() {
-                Value::String(hash) => hash.clone(),
-                _ => unreachable!()
-            }
-        }
-        _ => unreachable!()
+        Value::Object(m) => match m.get("hash").unwrap() {
+            Value::String(hash) => hash.clone(),
+            _ => unreachable!(),
+        },
+        _ => unreachable!(),
     };
 
     let mint_block = BlockHash::from_hash(Hash::from_str(&hash).unwrap());
@@ -119,7 +83,41 @@ fn test_new_blocks_with_mint_txs() {
     assert_eq!(txs1.last().unwrap().bag_id, [1; 32])
 }
 
-fn consensus_encode_tx<S: std::io::Write>(
+pub fn setup_bitcoin_node(port: u32, rpcport: u32, datadir: &str) -> std::process::Child {
+    let child = std::process::Command::new("bash")
+        .arg("./tests/setup_single_node.sh")
+        .arg(port.to_string())
+        .arg(rpcport.to_string())
+        .arg(datadir)
+        .stdout(Stdio::null())
+        .spawn()
+        .unwrap();
+    child
+}
+
+pub struct TempDir {
+    path: String,
+}
+
+impl TempDir {
+    pub fn new(path: String) -> Self {
+        std::fs::create_dir(path.as_str()).unwrap_or_else(|_| {
+            std::fs::remove_dir_all(path.as_str()).unwrap();
+            std::fs::create_dir(path.as_str()).unwrap();
+        });
+        Self { path }
+    }
+}
+
+impl Drop for TempDir {
+    fn drop(&mut self) {
+        std::fs::remove_dir_all(self.path.as_str()).unwrap_or_else(|_| ());
+    }
+}
+
+// `bitcoin` crate provide uparseable output with it's `consensus_encode` method for transactions
+// with zero inputs.
+pub fn consensus_encode_tx<S: std::io::Write>(
     tx: &Transaction,
     mut s: S,
 ) -> Result<usize, std::io::Error> {
